requirements.txt
absl-py
asgiref
astunparse
beautifulsoup4
blinker
certifi
charset-normalizer
click
colorama
deepface
dlib
face-recognition
Django
djangorestframework
filelock
fire
Flask
flatbuffers
gast
gdown
google-pasta
grpcio
gunicorn
h5py
idna
importlib_metadata
itsdangerous
Jinja2
keras
Keras-Preprocessing
libclang
Markdown
MarkupSafe
mtcnn
numpy
opencv-python
opt-einsum
pandas
pillow
protobuf
PySocks
python-dateutil
pytz
requests
retina-face
six
soupsieve
sqlparse
tensorboard
tensorboard-data-server
tensorflow
tensorflow-estimator
tensorflow-io-gcs-filesystem
termcolor
tqdm
typing_extensions
tzdata
urllib3
Werkzeug
wrapt
zipp